Bump the docker image version to 21
The increase in version was motivated by the need to fix
the failing RPM tests. Newer images contain the openssl-devel
and libffi-devel packages which fixed the RPM tests. In addition,
we had to correct the way we set the password for the app-admin
user in the centos-ssh-oj8 Dockerfile because the parent image
moved a configuration file.